requirements.gcp.txt
apache-airflow[google]==2.4.3
google-api-python-client 
google-auth-httplib2
google-auth-oauthlib
sqlalchemy-bigquery==1.4.4
apache-airflow-providers-google
apache-airflow-providers-http
requests
types-requests
jsonpath-ng
